+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 * The dispatcher offers two new parameters to control the client's call and stream timeout:
   - `call_timeout`: The maximum time to wait for a response from the client.
   - `stream_timeout`: The maximum time to wait before restarting a stream.
+* While the dispatch stream restarts we refresh our dispatch cache as well, to ensure we didn't miss any updates.
 
 ## Bug Fixes